rename condition -> condition_classes
