+++ droidlet/interpreter/condition_helper.py
@@ -4,7 +4,7 @@ Copyright (c) Facebook, Inc. and its affiliates.
 from typing import Optional, Union
 from word2number.w2n import word_to_num
 from droidlet.shared_data_structs import ErrorWithResponse
-from droidlet.interpreter.condition import (
+from droidlet.interpreter.condition_classes import (
     Condition,
     NeverCondition,
     AndCondition,
