+++ droidlet/interpreter/interpreter.py
@@ -7,8 +7,6 @@ import datetime
 from typing import Tuple, Dict, Any, Optional
 from droidlet.event import dispatch
 
-# FIXME agent
-from droidlet.dialog.dialogue_objects import DialogueObject, ConfirmTask
 from .interpreter_utils import SPEAKERLOOK
 # point target should be subinterpret, dance should be in agents subclassed interpreters
 from .reference_object_helpers import ReferenceObjectInterpreter, interpret_reference_object
@@ -17,7 +15,8 @@ from .filter_helper import FilterInterpreter
 from droidlet.shared_data_structs import ErrorWithResponse, NextDialogueStep
 from .task import maybe_task_list_to_control_block
 from droidlet.memory.memory_nodes import TripleNode, TaskNode
-
+# NOTE: this should be removed
+from droidlet.dialog.dialogue_objects import DialogueObject, ConfirmTask
 
 class Interpreter(DialogueObject):
     """
